+++ et-driver/et_vqueue.c
@@ -76,6 +76,7 @@ static void destroy_msg_node(struct et_msg_node *node)
 static void mm_reset_completion_callback(struct et_cqueue *cq,
 					 struct device_mgmt_rsp_hdr_t *rsp)
 {
+	int rv = 0;
 	struct et_pci_dev *et_dev = pci_get_drvdata(cq->vq_common->pdev);
 
 	if (rsp->status != DEV_OPS_API_MM_RESET_RESPONSE_COMPLETE) {
@@ -85,12 +86,12 @@ static void mm_reset_completion_callback(struct et_cqueue *cq,
 		return;
 	}
 
-	// TODO: SW-11288: The init timeout should be 0 secs here, currently
-	// MM reset response does not indicate the completion of reset.
-	if (et_ops_dev_init(et_dev, 60U) != 0) {
-		dev_err(&et_dev->pdev->dev, "Ops Device re-init failed!\n");
-		return;
-	}
+	et_ops_dev_destroy(et_dev, false);
+	rv = et_ops_dev_init(et_dev, 0, false);
+	if (rv)
+		dev_err(&et_dev->pdev->dev,
+			"Ops device re-initialization failed, errno: %d\n",
+			-rv);
 }
 
 void et_destroy_msg_list(struct et_cqueue *cq)